// src/app/interfaces/orderdetailinterface.ts
export interface Orderdetailinterface {
    // imagen principal del producto
    image: string;
    // nombre del producto o referencia
    prod_name: string;
    // descripción del producto
    prod_descriptions: string;
    // precio indivudual
    price: number;
    // cantidad seleccionada
    quantity: number;
}